+++ packages/api-derive/src/council/proposals.ts
@@ -2,12 +2,17 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { ApiInterfaceRx } from '@polkadot/api/types';
-import { DeriveCollectiveProposals } from '../types';
+import { Hash } from '@polkadot/types/interfaces';
+import { DeriveCollectiveProposal } from '../types';
 
 import { Observable } from 'rxjs';
-import { proposals as collectiveProposals } from '../collective';
+import { proposal as collectiveProposal, proposals as collectiveProposals } from '../collective';
 import { memo } from '../util';
 
-export function proposals (instanceId: string, api: ApiInterfaceRx): () => Observable<DeriveCollectiveProposals> {
+export function proposal (instanceId: string, api: ApiInterfaceRx): (hash: Hash | Uint8Array | string) => Observable<DeriveCollectiveProposal | null> {
+  return memo(instanceId, collectiveProposal(instanceId, api, 'council'));
+}
+
+export function proposals (instanceId: string, api: ApiInterfaceRx): () => Observable<DeriveCollectiveProposal[]> {
   return memo(instanceId, collectiveProposals(instanceId, api, 'council'));
 }